[svn] Added AI to CtD
Fairly simplistic, equates to "easy" mode if difficulty levels are ever
included.  First looks for any scoring moves, if none it will pick a random
safe move, and if no safe moves it will pick a random move with what is left.

Future (long range?) enhancement is to add look-ahead