[svn] Inserted checks on labels being NULL to avoid having GTK complain about failed assertions.
The problem is that the server sends text messages (player names, player
messages, etc) as soon as they become available, but the client does
not set up the table until it first receives a hand (at which time it
first has an idea of what size the table should be).  Thus, before the
first hand is sent any messages will be worthless _as far as the graphics
go_.  However, these messages are saved and used when the table is
set up, so there shouldn't be any problem with the current system
other than its lack of elegance.

An alternative would be to setup the table earlier, most likely as soon
as the client finds out how many players there are.  However, we'd have
to guess at the hand size and end up having to re-setup and re-size the
table later when the hand is sent.  This would be ugly for the user; it's
cleaner just to leave the splash screen up the whole time.

The only real drawback of the current system in this regard is if the
hand size changes.  The client remembers the "maximum hand size", which
it uses for sizing the table.  When the first hand is received, the
max hand size is set and the table is setup.  If a larger hand then
follows, the table will be set up again, most likely with a resize at
the same time.  There might even be strange games where we'd want to
resize the window to be smaller, not leaving it at the _maximum_ hand
size but basing it upon recent hand sizes.  Making these possible would
require additions to the protocol which I do not consider worthwhile,
so are likely to remain as they are.