[svn] Added support for the Sueca game, implemented by Ismael Orenstein.
I applied the patch submitted to SF as-is, except for some changes to the
ggzcards core code: indenting lines consistently, and fixing a small bug
in cards.c.  games/sueca.[ch] were added as-is.  There's one large bug in
the Sueca code itself which hasn't been addressed.

Thanks Perdig!  Soon the client will include you in the credits.