[svn] Added "Bid History" message.  It's a generalized message, maintained by the common code and controlled by game.bid_history (enabled by default).
Unlike the other messages, the code to set up this one is mostly in the
core code in common.c.  This is because even if game.bid_history is not
set, the bid history itself is still maintained - and can be used by
game-specific code to send out whatever message is desired.  The game
flag just controlls whether the message is sent by default.

For some games (like Hearts and Sueca) this should be disabled.