[svn] "Fixed" the bug that was causing the La Pocha clients to crash.  This was caused by a seg fault in the client when, on the _second_ time that the client had to increase it's hand size, free() was called to destroy the old hand.
Calling free() here wasn't a good idea to begin with - it should be
realloc, since it's possible that we need to save some of the hand data.

However, I have no clue why free() would fail on what seems to be a
perfectly normal g_malloc'd pointer.  To fix it, I have simply removed
the free() call - creating a memory leak.