[svn] Implemented rubber bridge!
Namely, there are three games to 100.  After one team wins 2 of these, the
rubber is over and they get a bonus.  However, we don't end the game here
(this is an option that will be added in the near future); rather, we
continue to play another rubber with a running tally of the scores.

There are just a few things left for this to be an excellent implementation
of Bridge:
	- Set up declarer properly; it should be the first person to
	  bid the suit rather than the most recent bidder of that suit.
	- Add bonus points for all the different things you get bonus
	  points for (vulnerability, slams, honors, etc.).