[svn] Changed the interface to the player message code to be much more powerful.
Now, instead of accessing the string directly, the game-specific code just
calls put_player_message() and add_player_message().  Much cleaner.

One down side is that the current implementation of the above functions
makes add_player_message very inefficient.  I dont think this will be a
problem, but if it is a more effecient implementation could be devised
without changing the interface.