[svn] "Fixed" the implementation of global messages so that the server can reliably track all of them.
I did this by making them into a linked list; not the most efficient
solution but definitely a step up.  A hash would be better, but I'm
not sure what hash library to link to (ggzd doesn't seem to use glib),
and we have several additional needs that hash libraries don't
always satisfy.