[svn] Fixed bug #435249: too-long chat messages causes disconnect.
Now ggzcore will not send more than the appropriate number of characters,
as determined by server->net->chat_size.  To do this without changing the
message buffer, a new buffer is created and the string is copied over
with snprintf (safely; there won't be any problem with %'s).

At a future time, it may be desirable to have the ggzcore code
automatically split up the message buffer and send multiple chat packets.
This would make things easier on the writers of client frontends (who
may wish to do this anyway), but would make it slightly easier for
malicious or stupid clients to flood the system.