[svn] Fixed things up so ggzmod should be installed as a shared library. 	- Added ggzmod/Makefile to configure.in. 	- Added ggzmod as a subdirectory to Makefile.am. 	- Chamged ggzmod/Makefile.am to compile/install libggzmod.la, 	  very similar to the easysock installation.
Note: previously, ggzmod/Makefile.am was compiling ggzmod as "libggz"
instead of "libggzmod".  I felt that this was definitely very
misleading, so to keep it in line with libggzcore I changed it.  The
same thing should be done to libggzdmod at the server end.  I left the
header file as "ggz_client.h" although "ggz_client_mod.h" or something
like that might be more appropriate.