[svn] Another small cleanup, this time in ggz_client_connect.  Manual null-termination of fd_name is not needed, since snprintf will assure that it is null-terminated.
Also, I added a comment about a potential bug.  Since fd_name is a fixed
length, in the (currently impossible) event that the file name for the
unix domain socket is too long the name will be cut short.  This will
result in ggz_client_connect silently failing and the game will most
likely hang.  The "right" way to do this would be to determine the string
length dynamically, but this would make the code slightly uglier.