[svn] Practically finished client-common pseudo-library.
It's not really a library, just three separate files: common.[ch] and
protocol.h.  The rest of code works off of this pseudo-library through
a system of function calls and callbacks.  The interface to the library
(in common.h) is documented using doxygen; see
http://devon.dhs.org/ggz/doxygen/.

I really don't know how to effectively split this code off into a real
library so that multiple clients can use it.  It should go into a ggz
module, but none fits.  Should a ggz-game-libs module be created?  Would
this be of use to other games?  The way it is now, updating it would be
a mess with multiple client frontends, and the code itself is just a big
jumble of functions in common.c.  I could partially split it off by
placing it in a separate directory within ggzcards, but this is an
irrevocable step (using CVS) so I won't do that just yet.

TODO: After reviewing this code, there are several points in it that need
to be updated/expanded on/cleaned up.  I don't remember where they all are
right now, but when I read this later I'll probably remember most of them.