[svn] Removed spurious enumerations and #defines from ggz_protocols.h.
It looks like the file was just an old copy of protocols.h from ggz, thus
it had all of the ggzd<->ggz client communication definitions in it as
well.  None of this is necessary or wise to duplicate, so I took out all
but the ggzdmod (ggzd<->game server) enumerations.