[svn] Added a "default event" to Perdig's event-based ggzdmod interface.
The internals are clumsy, but it should be easy enough from the game
programmer's point of view.  Doing
	ggzdmod_set_handler(GGZ_EVENT_DEFAULT, &handler_function);
will set handler_function() to handle all unhandled ggzdmod events (all
events that don't already have handlers).  For the typical game with a
single update() function (ugly!), this is ideal.

Note that the event-based system is in most cases much better than the
old manual system.  The GGZ main loop is moved into libggzdmod, which is
much cleaner from the game programmer's point of view!  I'm going to
convert GGZCards over straightaway...

The only problem I see is with real-time games.  They probably need their
own main loop so that they can do actions in real time (as opposed to
actions in response to ggzd or client messages).  This could be allowed by
adding another event, GGZ_EVENT_TIME (or something to that effect) that
is called periodically.  Of course, real-time games have lots of other
problems besides this one!