[svn] Changed player handler to not always return a gameover value.
Now it needs to be determined if it should ever return a gameover value.
For now it always return 0 except on error.  In fact, even the error
handling may be wrong - it returns -1 on any kind of error; when in fact
most errors probably aren't fatal.  Maybe it should just return 0.