[svn] Split off ggzdmod code into separate ggzdmod files.
  - ggzdmod_dispatch() handles input from the file descriptor.  This
    basically replaces the table_handle() function.
  - ggzdmod_send_launch, ggzdmod_req_gamejoin, and ggzdmod_req_gameleave
    send the appropriate packets across the socket.  (These names are
    rather poor; they should be changed in the future.)  I made some
    ugly code to set up the parameters for ggzdmod_send_launch.
  - Response packets (right now just RSP_GAME_OVER) are automatically
    sent by the ggzdmod code (after the handler function is called).
  - table_game_launch, table_game_join, table_game_leave, table_game_over,
    and table_log, and dbg_msg are used as pseudo-callbacks by the ggzdmod
    code.  This segregation should make it easy to split ggzdmod off into
    a separate library, if desired.  I had to change the parameters of the
    functions slightly (except dbg_msg).

Although this code looks good to me and GGZD runs well enough, I'm not
completely confident in its correctness.  Right now the behavior should
be completely unchanged, although with the ggzdmod code separated it will
be much easier to change the protocol.

Note that although the ggzd ggzdmod code is not a part of libggzdmod, it
is currently distributed in the same package (ggz-server), so compatibility
shouldn't be a problem so long as game servers use libggzdmod.