[svn] Implemented seat change request interface.
The game server calls ggzd_set_seat_status(seat, status) to request that
the seat type (assign) be changed.  ggzdmod then queries the server to see
if this is OK, and makes the change if so.  The server should then handle
the rest of the update.  The return value of the function indicates
success or failure.

Right now GGZD will always respond negatively to the request since it
can't yet make the change.

This should be an "extended" function that is not needed for normal
operation.  Perhaps it (and others like it) should be in a separate
file...