[svn] Separated some common client-server data into protocol.[ch].
This might be substantially easier to maintain as an outside game; all
modules could then be contained together within the same source tree and
it wouldn't be necessary to make sure they were all kept in sync.  On the
other hand, this would require work to set up configure to handle each
component and I really don't know how to do that.  A thought for another
day...