[svn] I implemented a VERY ugly checking system for the database that gives slightly better results:   - If no db is specified, db3 becomes the default.   - If anything other than db2/db3 is specified, a warning is printed and     db3 because the default.   - If db3 is to be used it is checked for.  If present it is used, if     not db2 is used.   - If db2 is to be used it is checked for.  If present it is used, if     not there is an error/exit.
Note that db3 is the default, but if it's not present db2 will be used
instead.  If db2 is specified, though, no check for db3 will be done on
failure.  I haven't been able to test these failure cases since that
would (I think) involve uninstalling db[23].

A much better system would have a "checkfordatabase" function that checks
for an arbitrary database.  The logic could then be more straightforward:
check for the given database, then each database in turn until one is
found or the list is exhausted.