[svn] Made many additional enhancements to spades bidding.
In counting ruff winners:
  - I entirely dropped the "trump sequence down from some high trump"
    p1 calculation.  There's no point in determining the maximum number
    of "guaranteed" trumps since this is just a subset of the spades
    strength calculation (handled separately).
  - Voids, singletons, and doubletons are calculated more sanely and
    are valued less highly.  Specifically, the calculation goes like this:
    * 90% chance of winning if you can trump the first round
    * 70% chance of winning if you can trump the second round
    * 40% chance of winning if you can trump the third round
  - I also dropped the extra length calculation, since it also overlaps
    with the spade strength calculation.
  - I questionably dropped the "exhausting all spades" penalty.  If you
    run out of spades ruffing tricks, that'll screw over your partner
    at the end of the hand.  On the other hand, this calculation somewhat
    overlaps with the weak-spade penalty (applied separately).

Spade strength winners are conted in a loop.  It's more elegant.

I made the set of parameters passed into these functions smaller,
calculating more information on-the-fly.

Counting non-spade tricks was made a little more conservative, with a
lower confidence level for unprotected high cards.  This section still
needs a lot of work though; there are a number of situations in which
it is quite inaccurate.

A separate find_final_bid() function was created.  The code therein
was simplified a little bit.

I created a function consider_bidding_nil() that returns TRUE if the
player should bid nil.  I reviewed all of the code in the function, and
changed a lot of it.  I also fixed a bad bug that had been in there since
I ported the code from NetSpades; it cause players to rarely bid nil.  I
think the calculated nil risk is on average larger now, but I'm not sure
if this is better or worse.

The actual get_bid() function is now quite short and understandable.

I introduced a preprocessor macro AGRESSIVE_BIDDING.  If set (not by
default) it will raise some of the constants used to calculate the
number of tricks.  But, the AI can't back up these more agressive bids
with good play.  Right now the bids are generally good conservative bids
that total to about 11.

Problems that remain:
  - Calculating non-spade tricks is inaccurate.
  - Calculating the nil risk is probably inaccurate.
  - The big thing: the AI is still very bad at playing.  In one demo hand,
    the bot got an incredibly good hand that it correctly bid 10 on, but
    was then completely unable to play the hand and only took 8.