[svn] Use the play count tracker to track the number of "count cards" each player has played.
"Count cards" are the ones that a player should have counted on to win a
trick; they can be used later for determining which cards remain in each
hand.  They include the A and K of non-trump suits, the A, K, Q of spades,
and any extra spade after the third.