[svn] Created a default Doxyfile for creating doxygen documentation.
To create the documentation, run "doxygen" within the top level directory.

Currently the files ggz_client.h and ggzcore.h are indexed; the
documentation created is for both ggzmod and ggzcore.  In the future this
could be changed.  It may also be wise later to have the documentation
itself be a part of CVS (either in this module or the docs module),
although it is about 90k in size.