[svn] Cleaned up ggzdmod's logging/debugging code.
I removed MSG_DBG, leaving only MSG_LOG in its place.  Of course, this
meant I had to remove all uses of MSG_DBG.  I also removed the log level
message from the ggzdmod communications; it is now up to ggzd to handle
this (ggzdmod-ggz calls a ggzd event callback function with the log info).

In summary, most of these changes are unnecessary, but will allow an
easier conversion to the new ggzdmod.