[svn] I've had enough of the fixed messaging system.  Although incredibly flexible, it's very limiting in terms of user interface issues.
So, I changed the protocol to allow different types of global messages
to be passed to the client.

Each message is still identified by a "mark" that is a single string.  In
the simplest case, the client can (and currently does) use this mark just
as the name of the message (for instance for the items on the "Messages"
menu).  In more complicated cases, the client may use this as an index
into its own system (ultimately this may be a reason for switching over to
an XML-style protocol, but that's a question for another day).

However, there are now different "types" of global messages.  The simplest
one is a "TEXT" message that is the same as the messaging system used now:
a single string, intended to be displayed as-is in a fixed-width font by
the client.  A second type of message is the "CARD GROUP" message, which
consists of a group of cards (one card per player).  A third is the "CARD
LIST" message, which consists of a list of cards (n cards per player).
(The card list message could be viewed as a superset of the card group
message, but since the display for the two is generally quite different
I've kept it like this for now.)  A final one is the "BLOCK" message,
which includes game-specific information that may or may not be used by
the client (in a format to be determined later).

For compatibility, I also bumped the protocol version number up.

It should be remembered that this is an experimental system; I'm really
going to have to see how things shake out as I go along.  The card group
and card list systems should be good enough to start with, once I put
together some code to use them.  After that, I'll see how things fall
together for the game-specific system.