[svn] Wrote send_last_hand_cardlist, with a wide variety of other changes to keep things clean.
The server will now send a cardlist global message for the "last hand"
message (in addition to the text message).

Also, the CARD GROUP message type has been removed (merged with CARD LIST),
and the format of the CARD LIST message has been changed so players can
have different numbers of cards in the list.