[svn] Fixed bug in comparing of old and new seat names, for real this time.
Each time we ggzdmod_set_seat, we free the old string and strdup the new
one.  This is fine, except if the new and old strings point to the same
places in memory.

Note that we're still using ggz_free/ggz_strdup, whereas elsewhere
malloc is used to allocate the string (in es_read_string_alloc).