[svn] Dammit!  Why can't I write code without bugs anymore!
  or

Fixed bug with ggzdmod_set_seat.  It _is_ allowed for games to change the
seat data, but each piece of data must be verified.  The only thing
allowed is to change the name of an AI player.  This cannot be checked
with strcmp() since one or both names might be NULL.