[svn] Extensive changes that allow more than one bid at a time.
It is now possible for more than one player to bid at a time.  The
game-specific code must just call the core bidding functions for each
player that is to bid, and make sure to request a bid from each player
that is to bid, and the core functions will be able to understand and
track the different players' bids.

This has not been tested.