[svn] Fixed small problem with handling of seat numbers in handle_join_event and handle_leave_event.
The problem is that players can join/leave before the game is initialized.
If this happens, the seat number of the player will be uninitialized; it'll
be set to -1.  In this case we can't mess with any of the seat data
(obviously).

This was already taken mostly taken care of, but I finished it up and made
some comments making it explicit.  Note that if a seat isn't properly
initialized on a join event, it may not get correctly initialized at all
later!