[svn] Fixed bug in ggzdmod.
seat_find_player() is used to search through a ggz_list to find a seat
with a matching name.  It uses strcmp() and (as ggz_list specifies, though
I find odd) returns 0 in the case of a match, 1 if there's no match.

But since seat names can be NULL, this will quickly segfault.  I changed
it so that it always returns 1 (no match) in this case.  This function is
used to find the seat that matches the name that ggzdmod-ggz sends when
a player leaves the table, so this should be the correct behavior.

I also added a comment which mentioned this in passing: ggzdmod-ggz should
_not_ send the name in this case.  It should send the seat number, or the
player's UID (although the game doesn't yet know the UID, so this isn't an
option yet).  But that's an issue for another day (soon).