[svn] Changed ggzdmod_connect to return 0 on success instead of the GGZdMod socket FD.
The reason for this is that games shouldn't be tempted to use this socket
at all - there's no reason for them (or ggzd, AFAICT) to need this
information.  I added a comment to this effect.

All current code just compares the return value to 0 (or ignores it
altogether), so this shouldn't conflict with anything.