[svn] Ugh.  Not only are the GGZSeatType enumeration values provided by ggzmod, they're provided by ggzcore as well.  This means that not only do they conflict with a game's local definitions (which is good), they also conflict with each other.
A temporary solution is to put each enumeration within an appropriate
#ifdef check, so that things will compile.  I'm not sure what the right
long-term solution would be.  I don't think it's right for games to be
encouraged to #include <ggzcore.h>, but the enumerations belong there
more than than do in ggzmod.  Perhaps another header file,
<ggz_client_common.h>, could be #included from both ggzcore.h and
ggzmod.h?  But unless there was more than one enumeration in the file,
it wouldn't be worth it.