[svn] Fixed things so that a bad easysock operation won't turn into a recursive loop.
The problem happened because I unthinkingly added a debugging message
(ggzdmod_log) in the case where there is an unhandled event.  But, if the
game doesn't handle the ERROR event (as none do right now), then a failed
connect will result in an ERROR event which is unhandled, and naturally
the ggzdmod_log call will give another ERROR event.  Of course, this
wouldn't be a problem if game servers handled this event, but...

Most game servers may have code like this:
  if (ggzdmod_connect(...) < 0) {
    fprintf(stderr, SOME_REALLY_LONG_ERROR_MESSAGE);
    exit(-1);
  }
would it be appropriate to have ggzdmod_connect write out this error
message instead?  This would deny the possibility of any translation, but
might otherwise be convenient for most game servers.