[svn] Moved GGZSeatType enumeration out of ggzmod.h and ggzcore.h, creating a new header file ggz_client.h that is included by both of them.
In theory there might be more code that can go into this file.  It may
eventually become worthwhile to move it into its own directory area.  For
now, though, I've just left it in the ggzcore directory and modified
ggzmod to link to it there.

Also, I remaned GGZ_SEAT_RESV to GGZ_SEAT_RESERVED to match what ggzcore
had for the enumeration.