[svn] Changed ggz_read_string_alloc to allocate memory with ggz_malloc.
This will break all code that currently uses this function, but is
nonetheless clearly the right thing to do - and should be done ASAP so
the calling code has more time to adjust.  IMO.

The internal implementation leaves a bit to be desired.  Previously if
the malloc failed we just returned an error, but ggz_malloc will (I think)
exit by default.  This is a bit of a security risk, as a malicious program
at the other end of the socket could send us lots of large strings and
perhaps force us to exit.  OTOH, we do have the alloc_limit so there is
some safety.