[svn] Unified ggzd's table player tracking data fields.
Old method: there was one char* array containing a "name" for each player.
This would be "<open>", "<reserved>", "<bot>", "<none>" for unfilled
(non-human) seats.  A separate array tracked the names of reserved
seats.

New method: there is a GGZSeatType array that tracks the types of each
seat.  A separate array tracks the names of reserved/human players.  It
should work well enough for spectators, too (if we implement them that
way).