[svn] On table launch, check and verify the names of any reserved players.
This fixes a security hole; the old code would overrun buffers or segfault
if given a NULL name.

We verify the name by looking it up in the player database.  If it is not
found, we just change the seat to OPEN.  Later this can be improved (as
discussed by Rich).