[svn] When a player joins, we must check to see if a seat (open or reserved) is avialable for them _before_ doing the TRANSIT_JOIN.
This means the work is doubled: when the transit is started, if there are
no open seats we must check each and every seat to see if it is reserved
_for that player_.  Later, when the transit_join is completed we must
first check all seats to see if they are a reserved seat for that player,
and if that fails check all seats to see if they are open (since we
always want to put a player in their reserved seat if possible).

With this change, reserved seats should pretty much work.  However, there
is one other case that has not been considered: if two seats are reserved
at a table for the same player.  ggzd does not check for this when the
table is lauched (to prevent it), and if someone does it, it will make it
very difficult to fill both seats (the player with the reservation would
have to join, leave, and rejoin).