[svn] Use strcasecmp instead of strcmp to compare player names in several places:   - When removing a player (by name) from a seat at a table.   - When seeing if a player matches a reservation.
The former should not be necessary (both the stored name and the given
name come straight from ggzd, so they should match).

The second should fix half of the seat reservation problem.  Now it should
be possible to reserve a seat for "crouton" and have "crouton", "Crouton",
or "CrOuToN" sit at that seat.  But if you try to reserve a seat for
"Crouton", the fact that the DB is done all-lowercase will still be a
problem, and the reservation will fail (currently it'll just be converted
to an OPEN seat).

Note, strcasecmp may not be available on all platforms, so we may need to
have configure check for it (?).