[svn] In table_check(), make sure the number of AI players is less than the total number of seats.
This will prevent players from launching all-AI games, which is a good
thing (at least for the moment).  This was a problem before with games
that would accept a wide variety of player #'s, like Chinese Checkers and
GGZCards.

Note, the user may try to launch such a game and will just get a cryptic
"Bad options" (E_BAD_OPTIONS) error from the server.  It may be
worthwhile to deal with this by checking the number of AI players
client-side before sending it on to the server (for instance gtk-client
could just disable the "AI" option for the first seat; ggzcore could
return an error if the client tried to send a bad player combination).