[svn] Switch the code to increase the maximum hand size back to using realloc instead of just mallocing the new hand and accepting a memory leak.  This used to cause a segfault from time to time, but now it should not.
In hindsight, it's obvious why things happened as they did.  Becasue the
buffer overrun on the hand was just 1 card (3 bytes), most of the time it
didn't affect anything and was untraceable.  Only for particular hand
sizes did things line up so that the overrun mangled malloc's internal
memory, and caused a crash.