[svn] Fixed errors in --disable-FEATURE options, for gettext and debug.
The format is AC_ARG_ENABLE(feature, descriptive_text, if_specified,
                            if_not_specified).
descriptive_text seems to have nothing to do with the rest.  If the
parameter is specified, we don't want to do anything - just let configure
it out.  If it's _not_ specified, then we need to give a default (which
we do).

The problem before was that there was an action given for if_specified -
which will lock us in to that action, no matter what is actually
specified.

Anyway, it should work now.