[svn] Workaround for bug.
When table_setup is called, we expect the game to have started, so there's
an assertion to that effect.  But, if we join a game in progress we get
no MSG_NEWGAME alert so game_started will not have been set.

The workaround is to change the assertion to instead set game_started and
log an error message.  But this should be fixed server-side.