[svn] Place the bid selection box directly onto the table, rather than in a separate dialog window - if the bid_on_table preference is set (turned off by default).
There may be problems with this system - for one thing, if there are too
many bids (like in Suaro and Bridge) they might not fit on the table at
all.  There may be drawing issues as well - we'll see.  This is why it's
not enabled by default, although it is much more convnient than the
separate window.