[svn] Changed the return type of a number of functions to be GGZPlayerHandlerStatus.
There may be more functions that return a value from this enum that I have
missed.

Also, there is at least one function (room_join) that may return either a
value from this enumeration _or_ an E_*** error code.  Currently this
works, but it could be dangerous (the sets _do_ overlap) and is logically
a bit confusing (returning GGZ_REQ_DISCONNECT means a disconnect, but
E_*** is the equivalent of GGZ_REQ_FAIL).  It would seem logical to
have the error code be sent in another parameter (or to set an errno-like
value), and return GGZ_REQ_FAIL in this case.