[svn] Completely rewrote the AI system so that the AI now runs client-side.
For each AI, the server will fork off an AI process that runs exactly as
a client does.  This has a number of advantages, mostly stemming from
encapsulation of the AI code - it is now completely self-contained; this
means its impossible for the AI to cheat even by accident (which I was a
bit worried about before), makes it easy to track separate state data
for each AI, makes it easy to use different AIs at the same time (not yet
supported), and makes it possible to keep the AI code cleaner overall.

There are only a few drawbacks.  The AI is a bit slower (although this
should not be noticable; certainly it'll still be faster than a human
player).  There's a bit more overhead for the server, which could be a
problem on a heavily loaded server (but again, less overhead than a human
player would case).  There are some extra open sockets, which could help a
heavily loaded server to reach it's _system_ socket limit (blah, blah,
blah).

The final (and biggest IMO) drawback is that the AI doesn't know anything
about the game state other than what is sent to the client - which is very
little.  For instance, the Suaro AI has to do some extra work to figure
out who the declarer is and what the contract is - work that is currently
duplicated between the server's game code and the AI code (but it may not
have to be).  More generally, an AI that begins playing in the middle of
a hand or a game (which is not currently possible, but eventually should
be) will not have full information about what has happened up to this
point in the hand.  Of course, this is no more problem than a human
would face in the same situation.

Finally, there are a number of limitations of the current implementation:
  - I've copied over the core client files (common.c/common.h) as
    client.c and client.h in the AI code.  This means an extra level of
    file duplication.
  - I had to simulate the work of ggzmod by calling socketpair/fork/exec
    for each AI process, and changing the core client files to not use
    ggzmod at all (that 3 is just hard-coded now).  This just means some
    extra work maintaining all of this stuff.  In the future, if ggzmod is
    changed to handle both GGZ and game side (just as ggzdmod does), the
    code could just use it directly (although this would introduce a
    dependency to ggz-client-libs).
  - If an AI process exits abnormally, there is currently no reasonable
    error handling.  Instead, the game will generally just hang.  This
    can be fixed in the future; the most likely handling would be to
    start a new "random" AI in place of the crashed one, and warn the
    players at the table.
  - We currently don't kill AI processes when the server exits.  But the
    AI code will exit on its own.  It would be easy enough to kill the
    process, but could lead to communication problems (we might try to
    send data to it, and end up starting a new AI in the process).  So
    care will be needed.
  - I had to rewrite the main loop of the server to call select() itself
    instead of using ggzdmod_loop().  This was a bit of a pain; it could
    be prevented by allowing ggzdmod to monitor _bot_ socket FD's, if they
    are set (and allowing the game to set them, of course).
  - It is extremely likely that bugs will turn up in this system.  This is
    a problem given the lack of error handling above.  But, the changes
    were so massive I really don't feel like reviewing every line of code
    in the patch.  The most likely source for bugs is that the AI's don't
    have access to all the information they used to.  I'd rather commit
    the changes now, and fix bugs as they come along.

Whew.