[svn] Add ugly-ass hack to keep ggz-gtk from segfaulting every time it starts up.
The new code to read the debug_types through ggzcore_conf_read_list was
extremely buggy.  That function returns the length of the list as an
integer.  However, debug_init expects the list to be null-terminated,
which it never was, hence...a near-guaranteed segfault on startup.

(This is the exact same issue as with command-line parameters for game
servers in ggzd, and the solution is the same: copy the list, and
null-terminate it.)