[svn] Put new library libgtk-games.a into directory common/.  Moved NetSpades' exit dialog into this library, and linked NetSpades to it.
The library now contains the ggz_show_exit_dlg() function.  In time it
can hold others as well, so that gtk-games all have a common feel.

I added a wrapper function to NetSpades so that the library function can
have a straightforward interface.  This is the only actual code change I
did; everything else was just moving code around.