[svn] Fixed most hastings problems.
The problem was that the game_send_move() function was supposed to send
the move to everyone, yet it returned if it ever encountered an AI player
or failed to send a move.

The fix is easy: just continue sending the move to the other players in
those cases.

However, there is still a problem, in that I'm pretty sure Hastings doesn't
check the move that is sent by the client.  This is not a showstopper,
however.