[svn] There was a buglet when we show the bid choices on the table: the label at the top of the bid selection table that says "Select your bid:" isn't drawn correctly when it's placed directly on the table.  (Justin tells me this is because a GtkLabel doesn't have its own GdkWindow.)
The workaround is to just not show that label when we have bid_on_table.
This is acceptable since the statusbar at the bottom of the table still
already says something similar.  It also has the advantage that more bids
can fit on the table area, so the bid_on_table option will work for more
games (in fact, it now works for suaro for me).