[svn] Fixed a major bug: initialize the AI's err_fd to -1, and check for an FD of -1 later when we're calling select() to wait for data on it.
This should prevent some Bad, highly unpredictable behavior from before
where the FD was initialized to 0 when the player array was allocated
(which happened during the GGZ launch), but not truly initialized to the
correct value until the AI was spawned (which doesn't happen until the
game is chosen).  This meant we could get some really fucked up
behavior when in the generic "GGZCards" room (in the other rooms, things
should work fine since we know the game type from the beginning).  In the
case of the kde client, the fact that the initial client packet from the
player (the one that sends the language) goes through must have made a
difference somehow, causing the server to lock up waiting for data on
FD 0.

In summary, it is likely that everything will work now, though more
stress-testing is needed.