[svn] Remove duplicate cards xpm's from La Pocha: it just uses the cards from the ggzcards directory directly.
I had lap use the ggzcards files instead of the other way around because
ggzcards uses additional rotated XPM files of the card fronts.  It is
possible the XPM files should go in the gtk-games common directory (?).

These XPM files are 500k in size, so removing the duplicate is a least
potentially a big saver in tarball size.  But a better solution would be
to have the images be converted to .png files and installed in a common
location - that way the installation/binary size could be kept down as
well.