[svn] Fix a rather hideous buffer overflow/non-termination problem in ggzd.
If a player logs in with a name longer than MAX_USER_NAME_LEN characters,
currently the length is unchecked.  In login_player() the name is just
copied over onto the database record with strcpy, and it's also passed in
to the hash code which does the same.

I've partially fixed this.  I fixed the strcpy() call in login_player() to
use snprintf as well.  But I have not fixed any other overflow problems
(there are still a lot of strcpy calls in ggzd, and most of them are
probably bad); rather, I just terminated the name manually after
MAX_USER_NAME_LEN characters in login_player().