[svn] Add the FD of the offending socket to the parameters of the socket error callback function
  -typedef void (*ggzIOError) (const char * msg, const GGZIOType type,
  +typedef void (*ggzIOError) (const char * msg,
  +                            const GGZIOType type,
  +                            const int fd,
                               const GGZDataType data);

and change the interal code to pass this parameter along with the others.
-1 is passed if there is no applicable FD.

I believe no code currently uses this callback except ggzcards (and that
not much, yet), so nothing should be disrupted by this change.  The
advantage is that the callee is now able to tell what socket the problem
occurred on, so action can be taken to correct it.  See ggz-dev for more.